+++ src/building/entertainment.c
@@ -15,7 +15,7 @@ static const building_type ENTERTAINMENT_BUILDINGS[] = {
 
 void building_entertainment_run_shows(void)
 {
-    for (int i = 0; i < NUM_ENTERTAINMENT_BUILDINGS; i++) {
+    for (size_t i = 0; i < NUM_ENTERTAINMENT_BUILDINGS; i++) {
         building_type type = ENTERTAINMENT_BUILDINGS[i];
         for (building *b = building_first_of_type(type); b; b = b->next_of_type) {
             if (building_monument_is_monument(b) && b->data.monument.phase != MONUMENT_FINISHED) {
